Move input stream adapter to its own header
Other input stream types are isolated to their own headers, this one should be as well.
