+++ src/leapserial/Archive.cpp
@@ -5,28 +5,6 @@
 
 using namespace leap;
 
-bool InputStreamAdapter::IsEof(void) const {
-  return is.eof();
-}
-
-std::streamsize InputStreamAdapter::Read(void* pBuf, std::streamsize ncb) {
-  is.read((char*)pBuf, ncb);
-  if (is.fail() && !is.eof())
-      return -1;
-  return is.gcount();
-}
-
-std::streamsize InputStreamAdapter::Skip(std::streamsize ncb) {
-  is.ignore(ncb);
-  if (is.fail() && !is.eof())
-    return -1;
-  return is.gcount();
-}
-
-bool OutputStreamAdapter::Write(const void* pBuf, std::streamsize ncb) {
-  return (bool)os.write((const char*)pBuf, ncb);
-}
-
 const char* leap::ToString(serial_atom atom) {
   switch (atom) {
   case serial_atom::ignored:
